Добавлено сквозное тестирование опроса датчиков. Исправлены сущности и ченьжлоги таблиц. Добавлен профиль тестирования slow для запуска медленных тестов. Добавлена проверка существования зоны перед запуском полива.

+++ src/main/java/ru/antizep/greenhouse/serial/command/PumpOnRequest.java
@@ -4,9 +4,9 @@ import java.util.Objects;
 
 public class PumpOnRequest implements ArduinoCommand {
 
-	private final int zoneId;
+	private final long zoneId;
 	
-	public PumpOnRequest(int zoneId) {
+	public PumpOnRequest(long zoneId) {
 		super();
 		this.zoneId = zoneId;
 	}
